Enhance test so it checks that the field is blank
Crux of the original issue was that when LDAP Users were created that they
weren't allowed to have blank first/last names (and this threw an an exception
when it happened).  When we pull the User from LDAP, lets actually check to
make sure that we not only got the User record out, but that the last name is
in fact what we expect it to be.