Use system logger instead of custom logger.
By hooking right into the system logger, we get the benefit of letting the
core communicate errors back to the User instead of our having to catch them
and then report them back to the User ourselves.  Also gives us the benefit of
having our log info captured and stored with the rest of the system logs.