Return "last good data" when LDAP refresh fails.
Addresses {bz: 2299}, "Bad LDAP Data Throws Validation Error".

When we are *refreshing* an existing LDAP User, if the data we retrieve from
LDAP fails to validate then use the last known good data for that User; we
obviously got that data from LDAP at _some_ previous point in time and it was
considered valid then, so treat it as valid now.

While we're at it, *DO* log a warning message to `nlw.log` indicating that
there has been a validation error, so its not completely silent.  Further,
*do* update the "last cached at" time for the LDAP User so that we don't keep
doing the lookup on this User on every single request/instantiation; if the
data is invalid now its not likely to change again before out cache TTL
expires so mark it as re-cached for performance reasons.

Unit tests added to verify that:
 * on initial auto-vivification of the LDAP User that data validation errors
   are still considered fatal
 * on _refresh_ of an LDAP User, data validation errors are logged and we then
   return the last cached data for that User.