Fix {bz: 2415} by short-circuit e-mail lookups.
When using `ST::User->Resolve()`, if the value passed in *looks like* an
e-mail address, we'll short-circuit and do a lookup *by* e-mail address
_first_.

Then, if we didn't find any matching User, we'll fall back to the regular
lookup/resolution mechanism.

This addresses Bug #2415, where a User wafl would get rendered with the wrong
User record details when multiple Users exist in the DB with the same e-mail
address, one of those Users being a "Default" user (as looking up by
'username' with the e-mail address would find the Default user first, even if
that isn't the actual active User record for that e-mail address based on the
configured user factories).