Deprecation warning in new 'st-refresh-ldap-users'.
To help ease users/support into using `st-ldap`, this script exists solely to
let people know that the old version of `st-refresh-ldap-users` has been
deprecated and that they should be using `st-ldap refresh-users` instead.

After displaying a warning to the user and pausing for 10 seconds, this script
automatically runs the new command.