Only look for wafl blocks at start of the string
The parser checks for each of the 12 different kinds of top level
blocks.  Previously, this code would try to match all of the top level
blocks, and then use the one that matched closest to the start of the
string.

This patch locks these top level block patterns to only match the start
of the string, and adds a new block 'else' as a last case.  This last
block matches any single line (blocks are always a full line).

This greatly speeds up pattern matching, b/c matching start of string is
very fast.