Erase LDAP config at start, so it doesn't pollute test.
This test suite is sensitive to the state of the LDAP configuration file and
what it contains (regardless of whether or not the listed LDAP servers are
reachable or not).

So, to make sure that we don't get any pollution from other LDAP tests that
may have run before us, erase the LDAP config file before starting things off.