Clear LDAP factory cache after the refresh is done.
Unit testing this material in-process identified that the factory cache that
we were using bled from one call to the next.  Although that shouldn't affect
things in production, it *did* affect things in the unit tests.

So, make sure that we clear the factory cache at the end of each refresh, so
it doesn't bleed from one call to the next.