Fetch all rows, instead of just one at a time.
For larger number of Users to refresh, this saves overhead on having to call
`$sth->fetchrow_arrayref()` for each User.

Stash pointed out that the current version of DBD::Pg grabs everything from
the DB when you execute the statement anyways, so calling fetchrow_arrayref()
to get rows one at a time is really just function call overhead.  Although it
might be easy to follow, DBD::Pg is going to do the bulk fetch of rows from
the DB on execute anyways so why not just grab them all from DBD::Pg in one
fell swoop and be done with it.