Disconnect the DBH before forking child processes.
stash pointed this out while we were discussing something else today.

Rather than disconnecting from the DBH at the start of a worker, it should
really be done *before* we even fork the child processes.  Otherwise we risk
running into possible race condition problems regarding the DBH and when it
gets closed.  e.g. if it gets closed by a child, it could close the DBH for
_all_ of the children (as Pg would have disconnected from the other end of the
DBH connection).

So, disconnect from the DBH at the start of the main loop, *before* we start
up any child processes.