Clear in-memory caches between jobs.
In case any of the in-memory caches are in use, make sure that they're cleared
between jobs.  That way we don't accidentally bleed information from one job
into the next, which may be stale by the time we get around to actually using
it in the second job.

e.g. if we sleep 60s between one job and another, there's no guarantee that
any of the information that we have in our cache is even going to be valid any
more, so be sure to clear the caches between jobs.

FYI, cache is cleared at the *end* of the job (rather than at the beginning of
the job) so that any cached handles or connections are closed down and aren't
left lingering while we sleep and wait for another job.