LDAP connection caching, which is *on* by default.
And which will be cleared automatically at the end of each HTTP request, when
`Socialtext::Handler::Cleanup` runs and clears out all of the in-memory
caches.

LDAP connection caching allows us to re-use a single LDAP connection for
multiple queries, rather than having to create a brand new LDAP connection
each and every time we want to talk to LDAP.

Much of time this isn't a big deal (as we're doing a minimal number of LDAP
lookups in a single HTTP request), but for those odd times when we need to do
a *lot* of LDAP lookups (e.g. loading users, refreshing stale LDAP User data)
not having a make a brand new connection to LDAP is beneficial.

Don't forget... not all of the perf issue with LDAP requests is on our end;
customers don't like it when you slam their LDAP server with dozens/hundreds
of requests while extracting User data from them.  They'd *much* rather prefer
to see a small handful of LDAP connections get opened and then have
dozens/hundreds of queries fired down each connection.