Cache YAML config files, saving us from re-parsing.
Cache the parsed YAML that we read in from the configuration file(s), so that
we don't have to *re-parse* it every single time we want to get at it.

Cache "freshness" check is the same as the one used in Socialtext::AppConfig;
if the last modified time is the same as we last saw _and_ the file size
hasn't changed, use the cached copy (otherwise, reload the file and re-prime
the cache).

Most notably, this saves us the overhead of re-parsing the YAML when we're
doing tight loops around LDAP User instantiation; instantiating an LDAP User
means needing to connect to LDAP, which in turn means having to load the LDAP
configuration from disk (which uses this code).