Fix default_workspace bugs.
If the default_workspace did not exist on the system and a user not yet
logged in hit the server root, they would get thrown into an infinite
loop. Now they are asked to log in.

After some discussion with PM, we decided the app should act as though
there is no default_workspace set if default_workspace's value does
not point to an existent workspace. This will result in less
disruption to the end users.