Be sure to load `ST::Cache`.
We use `ST::Cache` to help set up the LDAP Connection Cache, so be sure to
load it.  *Don't* expect or rely on any other module to load it on our behalf.

This fixes problems with `st-bootstrap-openldap` accumulating stale `slapd`
processes during unit test runs; the child process was dying when trying to
shut down OpenLDAP as it wasn't able to access the connection cache (as this
module wasn't loaded).  This was hidden/invisible, though, as we suppress
output from the child process as part of its daemonization.