Kinda crazy refactor of links for {bz: 2479}
I changed our old way of rendering links in HTML:

<a href="index.cgi?page_name">Some Label</a><!-- wiki-renamed-link Page Name -->

to

<a wiki_page="Page Name" href="index.cgi?page_name">Some Label</a>

This way, when you copy and paste the link, it doesn't lose the page name.
Before this change, you wouldn't copy the HTML comment, so the formatter
didn't know how to render the link and would change "hi"[ho] in to [hi].
Now, the additional attribute gets copied along with the link.

I also refactored a bunch of code for simplicity.