Make it so we can change locking ( and email_in ) without changing workspace settings.
Previously, if the admin was not changing the workspace permission set,
the code would not update the email_in or page locking status either.
Also, don't force an early exit if the perm set doesn't match one of our
predefined sets, we'll just call it 'custom' and move on.

Refactoring the code to make it a bit clearer what is happening.