Update the required schema version for this migration.
We require migration 59, not 57.

We also took some time to improve the way the actual migration
itself works. Before, we'd skip the migration if any
workspace_admin's had a lock permission, but since the goal
of the migration is to give all workspace_admin's lock
perms, we needed to rewrite the pre-check slightly.

In addition, we made the actual migration script not assume
that no workspace_admin's had a lock permssion, making the
migration a little more resiliant in the case of an upgrade
failure.