Simplify Apache config; these are equivalent.
When you're using a "custom_http_port", that *is* the main port for the HTTP
front-end; there's _no need_ to listen on both "the custom port" and "the main
port" and to redirect from one to the other... they're the *same* port!

Graham is sure that this got broken by accident at some point along the way,
but after conferring w/JohnT, Brandon, and PM, we're all of the opinion that
"if you were configuring a custom http port because you're not allowed to use
Port 80, then having a redirect from Port 80 to your custom port isn't going
to be of any use to you (so lets get rid of it)".

Use of a custom HTTP port *is* a supported configuration option, but no
redirect is provided from Port 80 to that custom HTTP port; you get *either*
"standard" or "custom", not both.