Make sure that we create a lock permission if needed.
Putting this in the pre-check seemed really wrong, so
we'll short circuit if the lock perm doesn't exist at that point,
it's impossible for any admins to have a permission that doesn't
exist.

Once in the actual migration, we add the lock perm if it doesn't
exist, then select it's ID ( getting currval wouldn't work becuase
it's possible that we didn't need to insert ) for use with granting
admins lock privs.