Rework the migration a little bit.
I really don't like the idea of updating the schema in a 'check' script,
so let's assume that if the schema isn't at the correct version, we
need to run the migration. Then we can let the script that is supposed
to change data update the schema.

This required adding a 'socialtext_schema_version' sub in
ST::Migration::Utils.