Redirect to "friendly name" when ssl_only enabled.
Rather than redirecting to the "%{SERVER_NAME}" Apache variable, *always*
redirect Users off to the name of the host as *we* think it is (which is the
hostname that we use in app-generated redirects, e-mail notifications, etc).

The Apache one _could_ be the "ugly name" for the server (which is the
fallback for our custom version anyways, but if "UseCanonicalName off" is set
in the Apache config its actually "whatever hostname was given in the 'Host'
HTTP header".

When doing SSL-only redirects in combination with a friendly web name, we
*always* want to redirect the User to the friendly name for the server.  So,
just do it that way in the Apache config.