FIX: allow for custom SSL port configuration
This apppears to have been broken as part of the refactoring that was done to
set up `Socialtext::HTTP::Ports` (as it never had support added to it for
custom SSL port configuration).

This commit changes the semantics of the configuration option slightly... if
you ask for the `ssl_port` directly you now get "0" meaning "no custom SSL
port" (so you should use the default port 443).  Previously, you'd explicitly
get back "443" as the default response.  Functionally, though, you should be
using `Socialtext::HTTP::Ports` to access information on which ports are
actually in use, which _will_ return you "443" if no custom port has been
configured in `socialtext.conf`.