Explicitly set custom SSL port for dev-envs.
When generating the configuration for a dev-env, explicitly list the custom
SSL port that we're using in the `socialtext.conf` configuration.

Although `ST::HTTP::Ports` would return us this value anyways in a dev-env, we
have the precedence set of listing the "custom_http_port" when its "not the
standard HTTP port" and we're continuing that here... this _isn't_ the
default/standard HTTPS port (443), so explicitly list it in the configuration
in the dev-env conf files so its *painfully* obvious that we're using a custom
port (and which one we're using).