Bug 2163: Don't listen on SSL port if not enabled.
If we aren't going to be using the SSL port for anything, don't even bother
listening on it.  Otherwise, people could connect to the SSL port and end up
with a weird error message because their browser thinks that its trying to
make an HTTPS connection but instead its just doing an HTTP connection.

Fixes Bug {bz: 2163}.