New "group_factories" configuration option.
Similar to "user_factories" except that it is a list of the drivers/factories
used for Group creation, not for User creation.  Although you will generally
want this to be the same as your "user_factories" configuration, it is a
separate option to allow for the possibility of having different sets of
factories (e.g. talk to 'this' LDAP server for User info, but use 'that' one
for Group info; allowing for load-balancing across the LDAP servers).