Fix cache issue with the Hub
The hub caches the preferences object for the current_user when
preferences_object is called. Unfortunately the cache is not cleared when
the current_user is changed (as it is when sending notification emails). So
now we check and if the current user has changed, we refresh the cache.

Oh, and the Hub.pm has no unit tests (well, it has 1 now). How sad is that?