Fix for notification email timestamp issues
Timezone plugin no longer cached the "now" time; if one is asking
for 'now' presumably one means now and not X minutes ago.

Also, the get_date method was implicitly altering the Socialtext::Date
object being passed to it. The method produces a text version of the
date/time passed to it, adjusted to the current user's time settings.
There is no reason to alter the passed time. The method has been updated
to work with a copy of the passed time.