When introspecting, march the inheritance tree, don't look *just* at the class we have.
If we've got an object that is in a _derived_ class, calling
`$meta->get_attribute()` is *only* going to check if that class/object has the
given attribute.  `$meta->find_attribute_by_name()`, however, will search the
inheritance tree to see if any of the parent classes have defined that
attribute.