Invert POD coverage test: test *everything* for POD coverage.
Our old POD coverage test only really worked if you actually _remembered_ to
update it with the names of the new modules that you wanted to include for POD
coverage testing.  More often than not, people (self included) would forget to
do this, and we'd never really be sure if we had reasonable POD coverage.

This update inverts the test so that we now find *ALL* of the modules under
`lib/` and test them _all_ for POD.  Everything that's known to be broken at
this point in time has been marked as a "TODO" so its not a _failing_ test
per-se, but it is noted as "we should either clean it up, or explicitly note
that its *not* going to have any POD in it".