Allow for POD coverage testing of a sub-set of modules.
If you want to test the POD coverage for a small set of modules, you can now
run:

  perl t/coding-standard/pod-coverage.t <list-of-modules>

and this test will run against *just* the modules that you list on the command
line (rather than going and finding the entire set of available modules and
testing them all).

This should make it a whole lot easier to test modules for POD coverage one at
a time, while we whittle down the number of TODOs in here.