Skip missing test files.
The STCI environment doesn't get cleaned out and started with a virgin build
on each run, so its possible for files to be left around from previous runs.

Normally that's not a big problem, *except* when the file that's left around
happens to be a test file.  Test files that are left around by accident end up
being run as part of the build and can interfere with our ability to hunt down
what the smallest test set is that'd generate failure.

Rather than dying a horrible death, skip the test entirely.