Fix {bz: 2658}, by ensuring the Username lookups are done correctly.
In cases when we have _multiple_ LDAP attributes that change and we need to do
a lookup by Username, we need to have the attribute mapped properly when we
call `ResolveId()`

One day, we'll just get rid of this being called both "driver_username" at the
DB level and "username" at the object level, and this won't be quite so much
of an issue any more.  For now, though, we've got an attribute with different
names at different levels and we need to watch out for it.