Merge UGRs on Group import, if Group already exists.
If the Group already exists, don't skip the Group or choke, but _merge_ the
UGRs in from the exported data into the existing Group.  Users that exist
already in the target Group get left as-is, but Users present in the exported
data that are missing are added to the Group.