Start of helper package to do User->Workspace Role calculations.
With the advent of "Groups" we now have multiple ways in which a User could
have a Role in a given Workspace; either directly via a UserWorkspaceRole, or
indirectly through a combination of UserGroupRoles and GroupWorkspaceRoles.

This helper package exists to try to help encapsulate the direct+indirect
calculations into a single place, rather than scattering the calculations in
multiples places throughout the system.