Extend TheSchwartz API w/ cancel_job&move_jobs_by.
Both require a funcname and uniqkey.  What "uniqkey" means when using multiple
database handles is a bit ambiguous; since it's not an index that spans
multiple servers, it's not really "uniq".  Socialtext is currently just using
one database so this should be acceptable for now.

Uses internal TheSchwartz::Moosified guts, but hey: stash is the maintainer.

* TODO: don't update grabbed_until during move_jobs_by.
* TODO: support table prefix.
* TODO: make cancel_job work with multiple databases.