Grandfather in old modules for pod coverage testing.
As per the discussion at the previous retro, all existing modules that are
failing the pod coverage tests are being grandfathered in as "TODO".

Any new modules that were created *after* the discussion are *NOT* being
grandfathered in; those modules require that the appropriate POD be written as
per the discussion.