Better diagnostic output for grandfathered/todo modules.
Dip a bit lower and do the POD Coverage testing ourself (rather than via
`Test::Pod::Coverage`), so we can output better diagnostic output for the
modules that we've grandfathered in.

Rather than (a) making them look perfectly clean, or (b) having gobs of
diagnostic output generated which clouds actual failures, we now output the
grandfathered modules so that they look clean *but* such that it includes a
list of the missing sections for those grandfathered modules.

Keeps things to one line so they don't clutter up the results, but continues
to make it clear that these modules could use improvement.