Rename "account_id" in Groups table to "primary_account_id".
This is actually what it is, so lets be explicit about it in the DB so as not
to create any additional confusion about how many Accounts a Group is allowed
to be associated with.

Although initially its likely to be "only one", we will be requiring Groups to
have access to multiple Accounts in the not-too-distant future.