*DON'T* run code to clear jobs at module load, use a "no-ceq-jobs" fixture instead.
Running the code when the module is loaded means that you have to make sure
that you've got a DB in place *before* you start the test.  That just gets
ugly as it requires you to create a BEGIN block to set up your test fixtures
_before_ you load `Test::Socialtext::Search`.

So.... take this out and replace it with a new "no-ceq-jobs" fixture that
functionally does the same thing, but which now operates as a test fixture.
As a fixture, it depends on the DB and thus it'll always be there when
`ST::Jobs` goes to clear out the job queue.