Lazy-load ST::Workspace and ST::Account, to make loading ST::User less expensive.
This should help make loading ST::User not take forever, by it needing to pull
in ST::Workspace (which in turn pulls in *LOTS* of things).

NOTE: This should *NOT* be taken to be a final answer on addressing this; it
acts as a short-term solution to the immediate issue, but there's a ton of
other refactoring that needs to be done to reduce inter-dependencies between
modules.