Reduce HTTP keep-alive max requests and timeout.
By reducing the keep-alive requests/timeout, we'll free up Apache processes
more frequently so that they can move on to handling other HTTP clients.

These numbers should still be high enough, though, for NTLM to continue to
function as needed:
* if we need to trigger NTLM authentication, chances of it occurring later on
  in an HTTP request stream are _low_
* customers enabling NTLM are doing so on internal networks, which don't
  generally suffer from the same latency/bandwidth issues that we see on the
  open Internet.

FYI: a future possible follow-up change to this would be to auto-detect
whether or not the appliance has been configured with NTLM (or not) and to
adjust the values further based on the results.  e.g. if you're using NTLM
keep numbers higher, and if no NTLM then reduce to default Apache keep-alive
requests/timeout