Make "Recent Changes" and "Watchlist" e-mails a bit prettier.
I got sick of seeing a long list of links with *no* indentation, which made it
hard to read the e-mails and sort out the "pages" from the "people"; row after
row of links with no indentation just gives a long line of blue links.

This version is at least a bit prettier, and provides similar indentation to
what we're using already in the text versions of these emails.