Squash account-people streams into one stream.
Turns out that we can't reasonbly partition the stream by account:
* two people can be members of the same two accounts, duplicate events result (one for each stream)
* assume three people (A, B, C) in a triangle account-membership relationship:

        A
       / \
      1   2
     /     \
    B---*---C

B tags C's profile.  Because A shares (different) accounts with B and C, A
should see that tag event.  However, if we assert that A, B, and C must share
a common account then A won't see the event.  The "*" account above can be any
number of common accounts between B and C.