Replace "ST::User->ByAccountId" tests with a fuller set of tests.
About to make changes under the hood here, and want to make sure that the
tests explicitly cover all of the various cases.  Although the previous tests
provided ample coverage on the code, it wasn't obvious from reading the tests
that the ordering tests were actually verifying the order properly (as the
results for all of the old tests expected the same result set) (if it gave
them back in the wrong order, you'd never know).